+++ src/index.js
@@ -1,76 +1,99 @@
-
-import AlgoliaSearch from './templates/algolia/index.vue'
-import AlgoliaDialog from './templates/algolia-dialog/index.vue'
-import ButtonBlockList from './templates/button-block-list/index.vue'
-import LinkButton from './templates/button/link.vue'
-import LinkButtonBlockList from './templates/button-block-list/link.vue'
-import Button from './templates/button/index.vue'
-import Sharer from './templates/sharer-block/index.vue'
-import Card from './templates/card/index.vue'
-import CardGridList from './templates/card-grid-list/index.vue'
-import CardGridListWithLoadMore from './templates/card-grid-list/withLoadMore.vue'
-import CardGridBlock from './templates/card-grid-block/index.vue'
-import FeaturedPost from './templates/featured-post/index.vue'
-import Header from './templates/header/index.vue'
-import HeaderLogoAzion from './templates/header/visualid/azion/index.vue'
-import HeaderNavigation from './templates/header/navigation-tabpanel/index.vue'
-import HeaderDialogAlgoliaSearch from './templates/header/dialog-algolia-search/index.vue'
-import MobileRightSidebar from './templates/header/mobile.vue'
-import Footer from './templates/footer/index.vue'
-import DropdownThemeSwitch from './templates/dropdown-theme-switch/index.vue'
-import DropdownLangSelector from './templates/dropdown-lang-selector/index.vue'
-import KeyboardKey from './templates/keyboard-key/index.vue'
-import NewsletterSingleLine from './templates/newsletter/single-line/index.vue'
-import NewsletterCard from './templates/newsletter/card/index.vue'
-import NewsletterWide from './templates/newsletter/wide/index.vue'
-import Overline from "./templates/overline/index.vue"
-import TitleFlexList from './templates/title-grid-list/index.vue'
-import TitleVerticalList from './templates/title-list/index.vue'
-import ArticleHero from "./templates/post/hero/index.vue"
-import ReadableContent from "./templates/post/readable-content/index.vue"
-import TabMenuSearch from "./templates/tab-menu/with-search.vue"
-import TabMenuSearchLink from "./templates/tab-menu/with-search-link.vue"
-import TagListBlock from "./templates/tag-block-list/index.vue"
-import ContentSection from "./templates/content-section-block/index.vue"
-import EmptyState from "./templates/empty-state-block/index.vue"
-import ErrorStateBlock from "./templates/error-state-block/index.vue"
-import ContentQuoteBlock from "./templates/content-quote-block/index.vue"
+import AlgoliaDialog from './templates/algolia-dialog/index.vue';
+import AlgoliaSearch from './templates/algolia/index.vue';
+import ArticleHero from './templates/post/hero/index.vue';
+import Button from './templates/button/index.vue';
+import ButtonBlockList from './templates/button-block-list/index.vue';
+import CardGridBlock from './templates/card-grid-block/index.vue';
+import CardGridList from './templates/card-grid-list/index.vue';
+import CardGridListWithLoadMore from './templates/card-grid-list/withLoadMore.vue';
+import ContentLogoBlock from './templates/content-logo-block/index.vue';
+import ContentQuoteBlock from './templates/content-quote-block/index.vue';
+import ContentSection from "./templates/content-section-block/index.vue";
+import DropdownLangSelector from './templates/dropdown-lang-selector/index.vue';
+import DropdownThemeSwitch from './templates/dropdown-theme-switch/index.vue';
+import EmptyState from './templates/empty-state-block/index.vue';
+import ErrorStateBlock from './templates/error-state-block/index.vue';
+import FeaturedPost from './templates/featured-post/index.vue';
+import Footer from './templates/footer/index.vue';
+import FormContactUs from './templates/forms/contact-us.vue';
+import GridBentoBlock from './templates/grid-bento-block/index.vue';
+import Header from './templates/header/index.vue';
+import HeaderDialogAlgoliaSearch from './templates/header/dialog-algolia-search/index.vue';
+import HeaderLogoAzion from './templates/header/visualid/azion/index.vue';
+import HeaderNavigation from './templates/header/navigation-tabpanel/index.vue';
+import HeroBlockBase from './templates/hero-block-base/index.vue';
+import KeyboardKey from './templates/keyboard-key/index.vue';
+import LinkButton from './templates/button/link.vue';
+import LinkButtonBlockList from './templates/button-block-list/link.vue';
+import MobileRightSidebar from './templates/header/mobile.vue';
+import NewsletterCard from './templates/newsletter/card/index.vue';
+import NewsletterSingleLine from './templates/newsletter/single-line/index.vue';
+import NewsletterWide from './templates/newsletter/wide/index.vue';
+import Overline from './templates/overline/index.vue';
+import ReadableContent from './templates/post/readable-content/index.vue';
+import Sharer from './templates/sharer-block/index.vue';
+import TabMenuSearch from './templates/tab-menu/with-search.vue';
+import TabMenuSearchLink from './templates/tab-menu/with-search-link.vue'
+import TagListBlock from './templates/tag-block-list/index.vue';
+import TitleFlexList from './templates/title-grid-list/index.vue';
+import TitleVerticalList from './templates/title-list/index.vue';
 
 export {
-  AlgoliaSearch,
-  ContentQuoteBlock,
   AlgoliaDialog,
+  AlgoliaSearch,
+  ArticleHero,
+
+  Button,
   ButtonBlockList,
-  Card,
+
+  CardGridBlock,
   CardGridListWithLoadMore,
   CardGridList,
-  Footer,
-  DropdownThemeSwitch,
+
+  ContentLogoBlock,
+  ContentQuoteBlock,
+  ContentSection,
+
   DropdownLangSelector,
-  CardGridBlock,
+  DropdownThemeSwitch,
+
+  EmptyState,
+  ErrorStateBlock,
+
+  FeaturedPost,
+  Footer,
+  FormContactUs,
+
+  GridBentoBlock,
+
   Header,
+  HeaderDialogAlgoliaSearch,
   HeaderLogoAzion,
   HeaderNavigation,
-  HeaderDialogAlgoliaSearch,
-  MobileRightSidebar,
-  FeaturedPost,
+
+  HeroBlockBase,
+
   KeyboardKey,
-  NewsletterSingleLine,
-  NewsletterWide,
-  TitleFlexList,
-  ArticleHero,
-  ReadableContent,
-  TitleVerticalList,
-  NewsletterCard,
-  Overline,
+
   LinkButton,
   LinkButtonBlockList,
+
+  MobileRightSidebar,
+
+  NewsletterCard,
+  NewsletterSingleLine,
+  NewsletterWide,
+
+  Overline,
+
+  ReadableContent,
+
   Sharer,
+
   TabMenuSearch,
-  Button,
-  TagListBlock,
-  EmptyState,
   TabMenuSearchLink,
-  ErrorStateBlock,
-  ContentSection
+  TagListBlock,
+  TitleFlexList,
+  TitleVerticalList,
 }
+