+++ src/templates/post/hero/index.vue
@@ -14,40 +14,51 @@
           <h1 itemprop="headline" class="text-3xl lg:text-4xl font-medium text-color">
             {{ title }}
           </h1>
-          <meta itemprop="alternateName" :content="description" />
-          <h2  itemprop="alternativeHeadline" class="font-normal text-xl lg:text-2xl text-color-secondary">
-            {{ description }}
-          </h2>
-        </div>
-        <p itemprop="datePublished" class="font-normal text-xs text-color-secondary">
-          {{ date }}
-        </p>
-        <div v-if="authors" class="flex flex-col md:flex-row gap-4 md:gap-8">
-          <template v-for="({avatar, title, role}, index) in authors" :key="index">
-            <div class="flex gap-3 md:max-w-56 items-center">
-              <div class="w-8 flex items-center">
-                <Avatar size="xlarge" :image="`${avatar}?ims=40x40`" :ariaLabel="`${title} - ${role}`" shape="circle" />
-              </div>
-              <div itemprop="author">
-                <p class="font-normal text-sm">
-                  {{ title }}
-                </p>
-                <p class="font-normal text-xs text-color-secondary">
-                  {{ role }}
-                </p>
-              </div>
-            </div>
+          <template v-if="description">
+            <meta itemprop="alternateName" :content="description" />
+            <h2 itemprop="alternativeHeadline" class="font-normal text-xl lg:text-2xl text-color-secondary">
+              {{ description }}
+            </h2>
           </template>
         </div>
+        <template v-if="date">
+          <p itemprop="datePublished" class="font-normal text-xs text-color-secondary">
+            {{ date }}
+          </p>
+        </template>
+        <template v-if="authors" >
+          <div class="flex flex-col md:flex-row gap-4 md:gap-8">
+            <template v-for="({avatar, title, role}, index) in authors" :key="index">
+              <div class="flex gap-3 md:max-w-56 items-center">
+                <div class="w-8 flex items-center">
+                  <Avatar size="xlarge" :image="`${avatar}?ims=40x40`" :ariaLabel="`${title} - ${role}`" shape="circle" />
+                </div>
+                <div itemprop="author">
+                  <p class="font-normal text-sm">
+                    {{ title }}
+                  </p>
+                  <p class="font-normal text-xs text-color-secondary">
+                    {{ role }}
+                  </p>
+                </div>
+              </div>
+            </template>
+          </div>
+        </template>
       </div>
       <div class="flex gap-4 md:gap-0 flex-wrap justify-between items-center">
-        <div class="flex flex-wrap gap-2" v-if="tagList" itemprop="keywords">
-          <template v-for="({ tag, tagLink }, index) in tagList" :key="index">
-            <a :href="tagLink" :title="tag" target="_blank" rel="noopener noreferrer">
+        <template v-if="tagList || tag">
+          <div class="flex flex-wrap gap-2" itemprop="keywords">
+            <template v-if="Array.isArray(tagList)" >
+              <a v-for="({ tag, tagLink }, index) in tagList" :key="index" :href="tagLink" :title="tag" target="_blank" rel="noopener noreferrer">
+                <Tag :value="tag" severity="info" />
+              </a>
+            </template>
+            <template v-else>
               <Tag :value="tag" severity="info" />
-            </a>
-          </template>
-        </div>
+            </template>
+          </div>
+        </template>
         <slot name="sharer"/>
       </div>
     </div>
@@ -65,7 +76,7 @@ import Tag from 'primevue/tag';
 
 const props = defineProps({
 date: {
-  required: true,
+  required: false,
   type: String
 },
 title: {
@@ -76,21 +87,21 @@ authors: {
   type: Object,
   required: false
 },
-postImage: {
-  url: String,
-  alt: String
-},
 breadcrumbItems: {
   type: Array,
   required: true
 },
 tagList: {
-  type: Array,
+  type: Array || String,
   required: false
 },
 description: {
   type: String,
-  required: true
+  required: false
+},
+tag: {
+  type: String,
+  required: false
 }
 })