+++ src/templates/post/readable-content/index.vue
@@ -1,5 +1,5 @@
 <template>
-  <article class="prose prose-lg max-w-full">
+  <article class="prose prose-lg max-w-full prose-headings:font-medium">
     <slot />
   </article>
 </template>